type(refactor): add crud controller, fix arrow fn in es6 classes(fix)
